Populate data using client library
This commit uses the new client library functionality to get data for the /records and /entries pages, in addition to metadata for both of these pages. For the /entries page, we also compute the differences between different versions of records for a given primary key so that these changes can be displayed to the user.