Update openregister-ruby to v0.2.1
This PR updates the openregister-ruby gem to v0.2.1 and stores the RegistersClient object as a class variable to ensure we get the benefits of the new caching in the client library. It also sets the cache duration to one hour - the period after which new data is allowed to be retrieved for a register.